[Feat] (domain) GetFeedInfoCase 추가
