+++ core/domain/build.gradle.kts
@@ -10,6 +10,8 @@ dependencies {
     api(projects.core.data)
     api(projects.core.model)
 
+    implementation(libs.javax.inject)
+
     testImplementation(libs.junit)
     androidTestImplementation(libs.androidx.junit)
     androidTestImplementation(libs.androidx.espresso.core)
