+++ core/domain/src/main/java/com/appleroid/core/domain/GetFeedInfoUseCase.kt
@@ -1,0 +1,4 @@
+package com.appleroid.core.domain
+import javax.inject.Inject
+
+class GetFeedInfoCase @Inject constructor()
