+++ feature/home/src/main/java/com/appleroid/feature/home/HomeScreen.kt
@@ -232,97 +232,6 @@ fun QuestionScreen(
                     )
                 }
             )
-
-            FeedCard(
-                modifier = modifier,
-                profileImageRes = R.drawable.ic_circle_test_2,
-                moreImageRes = R.drawable.ic_home_contents_more,
-                nickName = "이상훈",
-                mbti = "ENFJ",
-                time = "10시간 전",
-                feedTitle = "ENFJ가 최고야",
-                feedContent = "어떻게 생각해 ??",
-                buttonTitleList = mutableListOf("1도 공감 안돼", "응 맞아!"),
-                buttonPercentList = mutableListOf(86, 14),
-                boxRow = {
-                    Box(
-                        modifier = modifier
-                            .fillMaxWidth()
-                            .wrapContentHeight(),
-                        contentAlignment = Alignment.Center
-                    ) {
-
-                        DottedDivider()
-
-                        Row(
-                            modifier = modifier.fillMaxWidth(),
-                            verticalAlignment = Alignment.CenterVertically,
-                        ) {
-                            Image(
-                                painter = painterResource(R.drawable.ic_eclipse_start),
-                                contentDescription = "eclipse start image"
-                            )
-
-                            Spacer(modifier.weight(1F))
-
-                            Image(
-                                painter = painterResource(R.drawable.ic_eclipse_end),
-                                contentDescription = "eclipse end image"
-                            )
-                        }
-                    }
-                },
-                resultRow = {
-                    ImageWithTextBtn(
-                        modifier = modifier,
-                        selectedImageRes = R.drawable.ic_plus,
-                        unselectedImageRes = R.drawable.ic_plus,
-                        text = stringResource(R.string.mbti_result),
-                        interval = 2.dp,
-                        arrangement = Arrangement.End,
-                        onClick = {
-
-                        }
-                    )
-                },
-                snsRow = {
-                    ImageWithTextBtn(
-                        modifier = modifier.weight(weight = 1F, fill = false),
-                        selectedImageRes = R.drawable.ic_like,
-                        unselectedImageRes = R.drawable.ic_un_like,
-                        text = "243",
-                        interval = 8.dp,
-                        arrangement = Arrangement.Start,
-                        onClick = {
-
-                        }
-                    )
-
-                    ImageWithTextBtn(
-                        modifier = modifier.weight(weight = 1F, fill = false),
-                        selectedImageRes = R.drawable.ic_comment,
-                        unselectedImageRes = R.drawable.ic_comment,
-                        text = "20",
-                        interval = 8.dp,
-                        arrangement = Arrangement.Center,
-                        onClick = {
-
-                        }
-                    )
-
-                    ImageWithTextBtn(
-                        modifier = modifier.weight(weight = 1F, fill = false),
-                        selectedImageRes = R.drawable.ic_vote,
-                        unselectedImageRes = R.drawable.ic_vote,
-                        text = "1,304표",
-                        interval = 8.dp,
-                        arrangement = Arrangement.End,
-                        onClick = {
-
-                        }
-                    )
-                }
-            )
         }
     }
 }
