+++ gradle/libs.versions.toml
@@ -39,6 +39,7 @@ androidxMetrics = "1.0.0-beta01"
 androidxCoreSplashscreen = "1.0.1"
 
 [libraries]
+javax-inject = { module = "javax.inject:javax.inject", version = "1" }
 androidx-core-ktx = { group = "androidx.core", name = "core-ktx", version.ref = "coreKtx" }
 junit = { group = "junit", name = "junit", version.ref = "junit" }
 androidx-junit = { group = "androidx.test.ext", name = "junit", version.ref = "junitVersion" }
